requirements.txt
click==7.1.2
Jinja2==2.11.2
MarkupSafe==1.1.1
python-slugify==4.0.1
text-unidecode==1.3
